+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Refactored
+
+- Use Jekyll::Cache instead of poro cache.
+
 ## [0.6.0] - 2024-10-30
 
 ### Added